+++ packages/dsw-tdk/tests/test_cmd_list.py
@@ -7,28 +7,28 @@ from dsw.tdk import main
 @pytest.mark.vcr
 def test_list_simple(dsw_env: dict):
     runner = click.testing.CliRunner()
-    result = runner.invoke(main, args=['list'], env=dsw_env)
+    result = runner.invoke(main, args=['--no-config', 'list'], env=dsw_env)
     assert result.exit_code == 0
     assert 'dsw:questionnaire-report:2.8.0' in result.output
-    assert 'myorg:questionnaire-report:2.9.0' in result.output
+    assert 'organization:questionnaire-report:2.9.0' in result.output
 
 
 @pytest.mark.vcr
 def test_list_drafts_only(dsw_env: dict):
     runner = click.testing.CliRunner()
-    result = runner.invoke(main, args=['list', '-d'], env=dsw_env)
+    result = runner.invoke(main, args=['--no-config', 'list', '-d'], env=dsw_env)
     assert result.exit_code == 0
     assert 'dsw:questionnaire-report:2.8.0' not in result.output
-    assert 'myorg:questionnaire-report:2.9.0' in result.output
+    assert 'organization:questionnaire-report:2.9.0' in result.output
 
 
 @pytest.mark.vcr
 def test_list_released_only(dsw_env: dict):
     runner = click.testing.CliRunner()
-    result = runner.invoke(main, args=['list', '-r'], env=dsw_env)
+    result = runner.invoke(main, args=['--no-config', 'list', '-r'], env=dsw_env)
     assert result.exit_code == 0
     assert 'dsw:questionnaire-report:2.8.0' in result.output
-    assert 'myorg:questionnaire-report:2.9.0' not in result.output
+    assert 'organization:questionnaire-report:2.9.0' not in result.output
 
 
 @pytest.mark.vcr
@@ -38,7 +38,7 @@ def test_put_bad_token(dsw_api_url: str):
         'DSW_API_KEY': 'foo',
     }
     runner = click.testing.CliRunner()
-    result = runner.invoke(main, args=['list'], env=env_vars)
+    result = runner.invoke(main, args=['--no-config', 'list'], env=env_vars)
     assert result.exit_code == 1
 
 
@@ -49,5 +49,5 @@ def test_put_bad_url(dsw_api_key: str):
         'DSW_API_KEY': dsw_api_key,
     }
     runner = click.testing.CliRunner()
-    result = runner.invoke(main, args=['list'], env=env_vars)
+    result = runner.invoke(main, args=['--no-config', 'list'], env=env_vars)
     assert result.exit_code == 1